Lab1/requirements.txt
mlflow==2.14.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
